requirements.txt
python-dotenv
requests
tiktoken
langchain
langchain-community
langchain-core
langchain-openai
chromadb
pdfplumber
pymysql
sqlalchemy
pandas
openpyxl
openai
numpy
tqdm